fix issues and apply linting
Hi there,

I wanted to see this working but needed to fix some issues first.
At some places a float was used as the argument for the functions `setMaximumHeight`, `setMinimumHeight` and `setIconSize`.
These now need to be int's. So wrapped them in `int()`'s.
Furthermore I sorted all imports and applied linting using `ruff` and `black`.
I used a max line length to 100 characters.

+++ GUI/RibbonPane.py
@@ -1,10 +1,11 @@
 from PyQt5 import QtGui
 from PyQt5.Qt import Qt
-from PyQt5.QtWidgets import QWidget, QHBoxLayout, QVBoxLayout, QLabel, QGridLayout
+from PyQt5.QtWidgets import QGridLayout, QHBoxLayout, QLabel, QVBoxLayout, QWidget
+
 from GUI import gui_scale
 from GUI.StyleSheets import get_stylesheet
 
-__author__ = 'mamj'
+__author__ = "mamj"
 
 
 class RibbonPane(QWidget):
@@ -53,8 +54,8 @@ class RibbonPane(QWidget):
 class RibbonSeparator(QWidget):
     def __init__(self, parent):
         QWidget.__init__(self, parent)
-        self.setMinimumHeight(gui_scale() * 80)
-        self.setMaximumHeight(gui_scale() * 80)
+        self.setMinimumHeight(int(gui_scale() * 80))
+        self.setMaximumHeight(int(gui_scale() * 80))
         self.setMinimumWidth(1)
         self.setMaximumWidth(1)
         self.setLayout(QHBoxLayout())
